refactor(backend): Standardize module visibility (#249)
