+++ backend/src/models/mod.rs
@@ -1,8 +1,8 @@
 mod email_confirmation;
-pub mod flag_capture;
+mod flag_capture;
 pub mod task;
-pub mod team_invites;
+mod team_invites;
 mod team_status;
-pub mod teams;
+mod teams;
 pub mod user;
 mod user_roles;
