+++ backend/src/routes/admin/teams/mod.rs
@@ -2,7 +2,9 @@ mod clear_confirmation_code;
 mod delete;
 mod generate_confirmation_code;
 mod list;
-pub mod update;
+mod update;
+
+pub use update::UpdateTeamModel;
 
 pub fn config(config: &mut utoipa_actix_web::service_config::ServiceConfig) {
     config.service(list::list);
