+++ backend/src/routes/teams/invitations/mod.rs
@@ -1,5 +1,5 @@
-pub mod accept_invitation;
-pub mod my_invites;
+mod accept_invitation;
+mod my_invites;
 
 pub fn config(config: &mut utoipa_actix_web::service_config::ServiceConfig) {
     config.service(accept_invitation::accept_invitation);
